ECommerce
This code is a simple format for an e-commerce system. it adds customers to the customer's list, adds orde, has a product list, ships the orders and etc.
if you download all of the files and run the ECommerceUserInterface.java if you type in cuts, prods, newcust, order,.. it will ask you to give it some inputs and you will see how it runs.